# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  internals_handlers.ChatMessageRequest:
    properties:
      message:
        type: string
    type: object
  internals_handlers.CreateActivityRequest:
    properties:
      activity_type:
        type: string
      notes:
        type: string
      photo_url:
        type: string
    type: object
...
  internals_handlers.UpdateCarePlanRequest:
    properties:
      humidity_requirement:
        type: string
      light_requirement:
        type: string
      watering_amount:
        type: string
      watering_frequency_days:
        type: integer
      watering_method:
        type: string
      watering_tips:
        type: string
    type: object
...
  plantPal_internals_models.ActivityLog:
    properties:
      activity_type:
        $ref: '#/definitions/plantPal_internals_models.ActivityType'
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      logged_date:
        type: string
      notes:
        type: string
      photo_url:
        type: string
      plant:
        $ref: '#/definitions/plantPal_internals_models.Plant'
      plant_id:
        type: integer
      updatedAt:
        type: string
    type: object
  plantPal_internals_models.ActivityType:
    enum:
    - watered
    - fertilized
    - repotted
    - photo_node
    - milestone
    type: string
    x-enum-varnames:
    - ActivityWatered
    - ActivityFertilized
    - ActivityRepotted
    - ActivityPhotoNode
    - ActivityMilestone
  plantPal_internals_models.CarePlan:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      humidity_requirement:
        type: string
      id:
        type: integer
      light_requirement:
        type: string
      plant:
        $ref: '#/definitions/plantPal_internals_models.Plant'
      plant_id:
        type: integer
      updatedAt:
        type: string
      watering_amount:
        type: string
      watering_frequency_days:
        type: integer
      watering_method:
        type: string
      watering_tips:
        type: string
    type: object
  plantPal_internals_models.DifficultyType:
    enum:
    - easy
    - medium
    - hard
    type: string
    x-enum-varnames:
    - EasyDifficulty
    - MediumDifficulty
    - HardDifficulty
  plantPal_internals_models.GrowthMetric:
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      growth_rate_status:
        $ref: '#/definitions/plantPal_internals_models.GrowthRateStatus'
      height_cm:
        type: number
      id:
        type: integer
      plant:
        $ref: '#/definitions/plantPal_internals_models.Plant'
      plant_id:
        type: integer
      recorded_date:
        type: string
      updatedAt:
        type: string
    type: object
  plantPal_internals_models.GrowthRateStatus:
    enum:
    - slow
    - moderate
    - fast
    type: string
    x-enum-varnames:
    - SlowGrowthRate
    - ModerateGrowthRate
    - FastGrowthRate
  plantPal_internals_models.Notification:
    properties:
      createdAt:
        type: string
      daily_summary_enabled:
        type: boolean
      default_snooze_duration_minute:
        type: integer
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      notification_enabled:
        type: boolean
      preferred_notification_time:
        type: string
      sound_alert_enabled:
        type: boolean
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/plantPal_internals_models.User'
      user_id:
        type: integer
      vibration_enabled:
        type: boolean
    type: object
  plantPal_internals_models.Plant:
    properties:
      activity_logs:
        items:
          $ref: '#/definitions/plantPal_internals_models.ActivityLog'
        type: array
      care_plans:
        items:
          $ref: '#/definitions/plantPal_internals_models.CarePlan'
        type: array
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      growth_metrics:
        items:
          $ref: '#/definitions/plantPal_internals_models.GrowthMetric'
        type: array
      health_score:
        type: integer
      id:
        type: integer
      location:
        type: string
      nickname:
        type: string
      reminders:
        items:
          $ref: '#/definitions/plantPal_internals_models.Reminder'
        type: array
      scans:
        items:
          $ref: '#/definitions/plantPal_internals_models.Scan'
        type: array
      species:
        $ref: '#/definitions/plantPal_internals_models.Species'
      species_id:
        type: integer
      status:
        $ref: '#/definitions/plantPal_internals_models.PlantStatus'
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/plantPal_internals_models.User'
      user_id:
        type: integer
    type: object
  plantPal_internals_models.PlantStatus:
    enum:
    - good
    - needs_attention
    type: string
    x-enum-varnames:
    - GoodPlantStatus
    - NeedsAttentionStatus
  plantPal_internals_models.Reminder:
    properties:
      completed_at:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      is_completed:
        type: boolean
      plant:
        $ref: '#/definitions/plantPal_internals_models.Plant'
      plant_id:
        type: integer
      scheduled_time:
        type: string
      snooze_count:
        type: integer
      task_type:
        $ref: '#/definitions/plantPal_internals_models.TaskType'
      updatedAt:
        type: string
    type: object
  plantPal_internals_models.Scan:
    properties:
      ai_output_id:
        type: integer
      analysis_id:
        type: integer
      captured_image_url:
        type: string
      confidence_score:
        type: number
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      id:
        type: integer
      plant_id:
        type: integer
      retake:
        type: boolean
      selectedSymptoms:
        items:
          $ref: '#/definitions/plantPal_internals_models.Symptoms'
        type: array
      updatedAt:
        type: string
      user_id:
        type: integer
    type: object
  plantPal_internals_models.Species:
    properties:
      common_name:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      difficulty_level:
        $ref: '#/definitions/plantPal_internals_models.DifficultyType'
      family:
        type: string
      id:
        type: integer
      origin:
        type: string
      plants:
        items:
          $ref: '#/definitions/plantPal_internals_models.Plant'
        type: array
      scientific_name:
        type: string
      updatedAt:
        type: string
    type: object
  plantPal_internals_models.Symptoms:
    enum:
    - yellow_leaves
    - brown_spots
    - brown_steams
    - weathering_leaves
    type: string
    x-enum-varnames:
    - LeaveYellowSymptoms
    - BrownSpotSymptoms
    - LeaveBrownSymptoms
    - WeatheringSymptoms
  plantPal_internals_models.TaskType:
    enum:
    - water
    - fertilize
    - mist
    - rotate
    - repot
    type: string
    x-enum-varnames:
    - WaterTask
    - FertilizeTask
    - MistTask
    - RotateTask
    - RepotTask
  plantPal_internals_models.User:
    properties:
      care_streak_days:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      notifications:
        items:
          $ref: '#/definitions/plantPal_internals_models.Notification'
        type: array
      phone_number:
        type: string
      plants:
        items:
          $ref: '#/definitions/plantPal_internals_models.Plant'
        type: array
      total_journial_injuries:
        type: integer
      total_task_done:
        type: integer
      updatedAt:
        type: string
    type: object
  plantPal_internals_response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
host: plant-pal-api-ohhx.onrender.com
info:
  contact: {}
  description: API for PlantPal - a plant care and health monitoring application.
  title: PlantPal API
  version: "1.0"
paths:
  /diagnosis:
    post:
      consumes:
      - multipart/form-data
      description: Upload a plant image for AI-powered disease diagnosis and open
        a chat session
      parameters:
      - description: Plant image
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Start a plant diagnosis
      tags:
      - diagnosis
  /diagnosis/{session_id}:
    get:
      description: Retrieve the chat history for a diagnosis session
      parameters:
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get diagnosis chat history
      tags:
      - diagnosis
  /diagnosis/{session_id}/chat:
    post:
      consumes:
      - application/json
      description: Send a follow-up message in a diagnosis chat session
      parameters:
      - description: Session ID
        in: path
        name: session_id
        required: true
        type: string
      - description: Chat message
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internals_handlers.ChatMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Send a chat message
      tags:
      - diagnosis
...
  /plants/{id}/activities:
    get:
      description: Retrieve all activity logs for a specific plant
      parameters:
      - description: Plant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/plantPal_internals_models.ActivityLog'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get activity log
      tags:
      - activities
    post:
      consumes:
      - application/json
      description: Record a new activity for a plant (watered, fertilized, repotted,
        etc.)
      parameters:
      - description: Plant ID
        in: path
        name: id
        required: true
        type: string
      - description: Activity details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internals_handlers.CreateActivityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/plantPal_internals_models.ActivityLog'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log an activity
      tags:
      - activities
  /plants/{id}/care-plan:
    get:
      description: Retrieve the care plan for a specific plant
      parameters:
      - description: Plant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plantPal_internals_models.CarePlan'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get care plan
      tags:
      - care-plan
    put:
      consumes:
      - application/json
      description: Update the care plan for a specific plant
      parameters:
      - description: Plant ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internals_handlers.UpdateCarePlanRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/plantPal_internals_models.CarePlan'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/plantPal_internals_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update care plan
      tags:
      - care-plan
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"